+++ .eslintrc.js
@@ -9,7 +9,7 @@ module.exports = {
   overrides: [
     {
       files: ["*.js", "*.jsx"]
-    },
+    }
   ],
   parserOptions: {
     ecmaVersion: "latest",
@@ -20,24 +20,39 @@ module.exports = {
     "comma-dangle": [
       "error",
       {
-        arrays: "always-multiline",
+        arrays: "only-multiline",
         objects: "never",
         imports: "never",
         exports: "never",
         functions: "never"
-      },
+      }
     ],
-    indent: ["error", 2],
     "linebreak-style": ["error", "unix"],
     "multiline-ternary": 0,
+    "no-restricted-syntax": [
+      "error",
+      "ForInStatement",
+      "LabeledStatement",
+      "WithStatement"
+    ],
+    "react/destructuring-assignment": [1, "always"],
+    "react/jsx-props-no-spreading": 1,
     "react/prop-types": 0,
-    "semi": [2, "always"],
+    "react/no-unstable-nested-components": 0,
+    "react/function-component-definition": [
+      2,
+      {
+        namedComponents: ["arrow-function", "function-declaration"],
+        unnamedComponents: "arrow-function"
+      }
+    ],
+    semi: [2, "always"],
     "no-param-reassign": [
       "error",
       {
         props: true,
-        ignorePropertyModificationsFor: ["state"]
-      },
+        ignorePropertyModificationsFor: ["state", "accumulator"]
+      }
     ]
   },
   settings: {