+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-# 1.0.0
+# 1.1.0
++ Upgrade dependencies
++ Refactor tests
++ Upgrade dependencies
++ Add QRCode to enable MFA
++ Move QRCode to menu
+
+## 1.0.0
 + Upgrade dependencies + reorganize Docker
 + Add MSW during development
 + Refactor MSW tests